// projects/vellum-lib/src/lib/types/input.types.ts
export type InputSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl';
export type InputTheme = 'primary' | 'secondary';
export type InputVariant = 'basic' | 'outlined';
export type InputType =
  | 'text'
  | 'email'
  | 'password'
  | 'number'
  | 'tel'
  | 'url'
  | 'search'
  | 'date'
  | 'time'
  | 'datetime-local'
  | 'month'
  | 'week';